Move paper pengwin hooks into SeekerMod
