+++ src/main/java/fruitymod/CharacterMod.java
@@ -2,5 +2,5 @@ package fruitymod;
 
 import basemod.interfaces.*;
 
-public interface CharacterMod extends EditCharactersSubscriber, EditRelicsSubscriber, EditCardsSubscriber, EditStringsSubscriber, EditKeywordsSubscriber {
+public interface CharacterMod extends EditCharactersSubscriber, EditRelicsSubscriber, EditCardsSubscriber, EditStringsSubscriber, EditKeywordsSubscriber, OnCardUseSubscriber, OnPowersModifiedSubscriber, PostBattleSubscriber, PostDungeonInitializeSubscriber {
 }
